scrollbar for card view set random name folder structure changed
